Removing MIT (commercial) license until I switch to only using OurAirports data since I belive this violates OpenFlights licensing (see description).
"The OpenFlights Airport, Airline, Plane and Route Databases are made available under the Open Database License. Any rights in individual contents of the database are licensed under the Database Contents License. In short, these mean that you are welcome to use the data as you wish, if and only if you both acknowledge the source and and license any derived works made available to the public with a free license as well.

Licenses for commercial use of data or derived images are available on request. Most simple cases, for example use of an image within a book or other printed publication, are granted for a flat fee of US$100. Please contact us for details.

Airport data derived OurAirports and DAFIF, as well as route data from Airline Route Mapper, is in the public domain. Airline and plane data derived from Wikipedia may be subject to the GNU Free Documentation License. Whether these databases pass the threshold of originality and are thus copyrightable in the United States is an open question, and OpenFlights does not assert the validity or lack thereof of such a claim."